@@ -25,16 +25,16 @@
     "firefox esr"
   ],
   "dependencies": {
-    "@angular/animations": "~17.3.12",
+    "@angular/animations": "~18.2.14",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "~17.3.12",
-    "@angular/compiler": "~17.3.12",
-    "@angular/core": "~17.3.12",
-    "@angular/forms": "~17.3.12",
+    "@angular/common": "~18.2.14",
+    "@angular/compiler": "~18.2.14",
+    "@angular/core": "~18.2.14",
+    "@angular/forms": "~18.2.14",
     "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "~17.3.12",
-    "@angular/platform-browser-dynamic": "~17.3.12",
-    "@angular/router": "~17.3.12",
+    "@angular/platform-browser": "~18.2.14",
+    "@angular/platform-browser-dynamic": "~18.2.14",
+    "@angular/router": "~18.2.14",
     "@types/semver": "^7.5.8",
     "@types/ua-parser-js": "^0.7.36",
     "@iqb/responses": "3.6.0",
@@ -52,10 +52,10 @@
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "^17.0.2",
-    "@angular-devkit/build-angular": "~17.3.17",
-    "@angular/cli": "^17.3.17",
-    "@angular/compiler-cli": "~17.3.12",
-    "@angular/language-service": "~17.3.12",
+    "@angular-devkit/build-angular": "~18.2.21",
+    "@angular/cli": "^18.2.21",
+    "@angular/compiler-cli": "~18.2.14",
+    "@angular/language-service": "~18.2.14",
     "@compodoc/compodoc": "^1.1.19",
     "@istanbuljs/nyc-config-typescript": "1.0.2",
     "@types/browserslist-useragent": "^3.0.4",
