@@ -17,7 +17,7 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { Pipe } from '@angular/core';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { AlertComponent, CustomtextPipe, MainDataService } from '../shared/shared.module';
 import { GroupMonitorComponent } from './group-monitor.component';
 import {
@@ -36,6 +36,7 @@ import {
 import { TemplateContextDirective } from '../shared/directives/template-context.directive';
 import { TimeLeftPipe } from './test-session/timeleft.pipe';
 import { PositionPipe } from './test-session/position.pipe';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 class MockMatDialog {
   open(): { afterClosed: () => Observable<{ action: boolean }> } {
@@ -110,16 +111,16 @@ describe('GroupMonitorComponent', () => {
         MatRadioModule,
         MatCheckboxModule,
         MatTableModule,
-        MatSlideToggleModule,
-        HttpClientTestingModule
-      ],
-      providers: [
+        MatSlideToggleModule],
+    providers: [
         { provide: TestSessionManager, useValue: new MockTestSessionManagerService() },
         { provide: MatDialog, useValue: new MockMatDialog() },
         { provide: BackendService, useValue: new MockBackendService() },
-        { provide: MainDataService, useValue: new MockMainDataService() }
-      ]
-    })
+        { provide: MainDataService, useValue: new MockMainDataService() },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+    ]
+})
       .compileComponents();
   }));
 
