@@ -1,10 +1,11 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { MatCardModule } from '@angular/material/card';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { Observable, of } from 'rxjs';
 import { WelcomeComponent } from './welcome.component';
 import { BackendService } from '../backend.service';
 import { ServerTime } from '../sys-check.interfaces';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 class MockBackendService {
   // eslint-disable-next-line class-methods-use-this
@@ -24,18 +25,17 @@ describe('WelcomeComponent', () => {
     TestBed.configureTestingModule({
       declarations: [
         WelcomeComponent
-      ],
-      imports: [
-        MatCardModule,
-        HttpClientTestingModule
-      ],
-      providers: [
+    ],
+    imports: [MatCardModule],
+    providers: [
         {
-          provide: BackendService,
-          useClass: MockBackendService
-        }
-      ]
-    })
+            provide: BackendService,
+            useClass: MockBackendService
+        },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+    ]
+})
       .compileComponents();
   }));
 
